Simplify and fix issue documentation
This commit updates build-content script (invoked during docker build)
to write all rules to simply
`./content/{#rule}.txt`

instead of nesting them under:

`./content/#{category}/#{rule}.txt`

The update is cleaner and fixes a problem:
1) no complexity content due to phpmd quirky discrepancy between rule
category name in documentation and in phpmd code itself.

`!tldr`:
My first approach tried to make the doc path during
build  match the check-naming in phpmd code, but this was messy and required
conditional logic in a few places.

The rule names are unique so they suffice for the lookup.

https://github.com/codeclimate/codeclimate-phpmd/blob/master/Category.php#L72