Include separator in directory expansion check
The intent of this function is to expand include_paths to individual
files and add each as a unit of work to the forking server. This gives
the maximum amount of concurrency given the server's configuration.

The check to see if the include_path was a directory prepended "/code",
but without a separator. For example, "/coodapp/lib/" instead of
"/code/app/lib". The check would return false and the entire directory
would be added as a single unit of work. This meant we achieved almost
no concurrency and customer's with large repositories would time out.