Add category for check new in 2.3
Since the update to PHPMD 2.3, we see invalid index warnings since this
check is not in our category map. The warnings don't cause errors and we
fall through to a default Style category and remediation, but these are
more appropriate values.

The values were chosen as the same as the too-many-methods check, by the
reasoning that the general task of reducing the size of objects is the
same remediation no matter the methods' visibilities. Note: we don't
currently have the facility to scale remediation by overage, which would
make sense in this case, but adding that has been deferred for now.

http://phpmd.org/rules/codesize.html#toomanypublicmethods