Don't be silly
Something is either intentionally changed or broken in edge when
resolving dependencies, readline and libxml2 are not installed, causing
php to segfault.

Building from edge is dangerous for exactly these reasons. This patch
fixes the tag to 3.3 (the latest released version of alpine at this
time), which still builds correctly.